+++ system/shaders/GLES/2.0/gles_shader_multi_blendcolor.frag
@@ -25,7 +25,7 @@ uniform sampler2D m_samp0;
 uniform sampler2D m_samp1;
 varying vec4 m_cord0;
 varying vec4 m_cord1;
-uniform lowp vec4 m_unicol;
+uniform highp vec4 m_unicol;
 
 void main ()
 {
